+++ src/lab/experiments/gamev2/GameRenderer.tsx
@@ -2,6 +2,7 @@ import React, { useRef, useState, useEffect } from "react";
 import { eventBus } from "./EventBus";
 import { HexCoordinate, createHexCoordinate, getNeighbors } from "./types/HexCoordinate";
 import { UnitData, initialUnits } from "./types/UnitData";
+import { HexCell } from "./components/HexCell";
 
 // Types
 interface GameRendererProps {
@@ -9,84 +10,10 @@ interface GameRendererProps {
   height: number;
 }
 
-interface HexCellProps {
-  coordinate: HexCoordinate;
-  unit?: UnitData;
-  isMoveable?: boolean;
-  onHover: (coord: HexCoordinate, isHovering: boolean, isUnit: boolean) => void;
-}
-
 // Grid constants
 const GRID = {
   WIDTH: 100,
   ROW_OFFSET: 50
-};
-
-// Hex Cell Component
-const HexCell: React.FC<HexCellProps> = ({ 
-  coordinate, 
-  unit, 
-  isMoveable,
-  onHover 
-}) => {
-  const [isHovered, setIsHovered] = useState(false);
-
-  const handleMouseEnter = () => {
-    setIsHovered(true);
-    onHover(coordinate, true, !!unit);
-  };
-
-  const handleMouseLeave = () => {
-    setIsHovered(false);
-    onHover(coordinate, false, !!unit);
-  };
-
-  const handleClick = (e: React.MouseEvent) => {
-    e.preventDefault();
-    
-    if (e.button === 0) {
-      if (unit) {
-        console.log(`Selected unit at (${coordinate.x}, ${coordinate.y}, ${coordinate.z}) with movement ${unit.movement}`);
-        eventBus.emit('unit-selected', { unitId: unit.id, position: coordinate });
-      } else {
-        console.log(`Clicked coordinate: (${coordinate.x}, ${coordinate.y}, ${coordinate.z})`);
-      }
-    } else if (e.button === 2) {
-      console.log('menu');
-    }
-  };
-
-  return (
-    <div 
-      style={{
-        width: `${GRID.WIDTH}px`,
-        height: `${GRID.WIDTH}px`,
-        backgroundColor: isHovered ? '#4a90e2' : 
-                        isMoveable ? '#98fb98' :  // Light green for moveable cells
-                        unit ? '#ffeb3b' : '#f0f0f0',
-        clipPath: 'polygon(0% 25%, 0% 75%, 50% 100%, 100% 75%, 100% 25%, 50% 0%)',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        cursor: 'default',
-        userSelect: 'none',
-        fontSize: '12px',
-        margin: 0,
-        padding: 0,
-        boxSizing: 'border-box',
-        flexShrink: 0,
-        flexGrow: 0,
-        position: 'relative',
-        transition: 'background-color 0.2s ease',
-      }}
-      onMouseEnter={handleMouseEnter}
-      onMouseLeave={handleMouseLeave}
-      onClick={handleClick}
-      onContextMenu={handleClick}
-    >
-      {unit ? 'U' : `(${coordinate.x},${coordinate.y},${coordinate.z})`}
-    </div>
-  );
 };
 
 // Main Renderer Component
